omega_bot/requirements.txt
# Core dependencies
python-telegram-bot>=20.0
diffusers>=0.24.0
transformers>=4.36.0
torch>=2.1.0
Pillow>=10.0.0

# Configuration and data handling
PyYAML>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
numpy>=1.24.0

# Development dependencies
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
